Fix some unit tests
Reviewed-on: http://gerrit.dmdirc.com/390
Tested-by: Gregory Holmes <greboid@dmdirc.com>

+++ test/com/dmdirc/config/ConfigSourceTest.java
@@ -22,7 +22,6 @@
 package com.dmdirc.config;
 
 import com.dmdirc.harness.TestConfigSource;
-import java.awt.Color;
 import org.junit.Test;
 import static org.junit.Assert.*;
 
@@ -38,7 +37,7 @@ public class ConfigSourceTest {
     
     @Test
     public void testGetInt() {
-        assertEquals(42, s.getOptionInt("true", "42"));
+        assertEquals(42, s.getOptionInt("true", "42").intValue());
     }
     
     @Test
